chore(server): move files
